split shmipc and libchronicle, refactor callbacks
